requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
click==8.1.7
fastapi==0.110.0
h11==0.14.0
idna==3.6
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
redis==5.0.3
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0
